==> pbmc_TBNK_comparisons_asapseq/code/00_HTO.R <==
library(data.table)
library(Seurat)
library(Matrix)
library(dplyr)
library(BuenColors)

import_kite_counts <- function(path){
  mtx <- fread(paste0(path,"/featurecounts.mtx"), header = FALSE)
  dim <- mtx[1,]
  mtx <- mtx[-1,]
  matx <- sparseMatrix(i = mtx[[1]], j = mtx[[2]], x = mtx[[3]])
  rownames(matx) <- fread(paste0(path,"/featurecounts.barcodes.txt"), header = FALSE)[[1]]
  colnames(matx) <- (fread(paste0(path,"/featurecounts.genes.txt"), header = FALSE)[[1]])
  return(t(matx))
}

ASAPC <- import_kite_counts("../data/three_comparison/HTO_ASAP_C")
sc <- fread("../data/three_comparison/broad_pbmcs_aggr_singlecell.csv.gz") %>%
  filter(cell_id != "None")
barcodes <- sc[[1]]
channel <- substr(barcodes, 18, 18)
cells <- gsub("-3", "", barcodes[channel == "3"])
cmat <- ASAPC[10:13,intersect(colnames(ASAPC), cells)]
dim(cmat)

pbmc.hashtag <- CreateSeuratObject(counts = cmat, assay = "HTO")
pbmc.hashtag <- NormalizeData(pbmc.hashtag, assay = "HTO", normalization.method = "CLR")
pbmc.hashtag <- HTODemux(pbmc.hashtag, assay = "HTO", positive.quantile = 0.9995)
table(pbmc.hashtag$HTO_classification.global)

Idents(pbmc.hashtag) <- "HTO_maxID"
pR <- RidgePlot(pbmc.hashtag, assay = "HTO", 
                features = rownames(pbmc.hashtag[["HTO"]])[1:4], ncol = 4, cols = jdb_palette("corona"))
#ggsave(pR, file = "plots/HTO_ridges.pdf", width = 10, height = 2.5)  

FeatureScatter(pbmc.hashtag, feature1 = "hto_HTO-1", feature2 = "hto_HTO-2")

# Visualize
Idents(pbmc.hashtag) <- "HTO_classification.global"
pbmc.hashtag.subset <- subset(pbmc.hashtag, idents = "Negative", invert = TRUE)

# Calculate a distance matrix using HTO
hto.dist.mtx <- as.matrix(dist(t(GetAssayData(object = pbmc.hashtag, assay = "HTO"))))

# Calculate tSNE embeddings with a distance matrix
pbmc.hashtag<- RunTSNE(pbmc.hashtag, distance.matrix = hto.dist.mtx, perplexity = 100)
DimPlot(pbmc.hashtag)
pbmc.hashtag <- ScaleData(pbmc.hashtag)
m <- t(pbmc.hashtag@assays$HTO@counts)
idx_mat <- model.matrix(~0 + as.factor(max.col(m)))
pbmc.hashtag@meta.data$pct_max_counts <- rowSums(idx_mat * m)/rowSums(m)
pbmc.hashtag@meta.data$pct_max_counts_g0.5 <- pbmc.hashtag@meta.data$pct_max_counts > 0.5
FeaturePlot(pbmc.hashtag, features = c("pct_max_counts_g0.5"))

pBase <- DimPlot(pbmc.hashtag.subset, pt.size = 0.01) +
  labs(x = "t-SNE1", y = "t-SNE2") +
  scale_color_manual(values = c("red2", "black")) + 
  theme_void() + theme(legend.position = "none") + ggtitle("")

ggsave(pBase, file = "../plots/doublets_tSNE.png", dpi = 500, width = 2, height = 2)


p1 <- FeaturePlot(pbmc.hashtag.subset, "HTO-1", max.cutoff = 'q99', pt.size = 0.01) +
  labs(x = "t-SNE1", y = "t-SNE2") + theme_void() + theme(legend.position = "none") + ggtitle("")

p2 <- FeaturePlot(pbmc.hashtag.subset, "HTO-2", max.cutoff = 'q99', pt.size = 0.01) +
  labs(x = "t-SNE1", y = "t-SNE2") + theme_void() + theme(legend.position = "none") + ggtitle("")

p3 <- FeaturePlot(pbmc.hashtag.subset, "HTO-3", max.cutoff = 'q99', pt.size = 0.01) +
  labs(x = "t-SNE1", y = "t-SNE2") + theme_void() + theme(legend.position = "none") + ggtitle("")

p4 <- FeaturePlot(pbmc.hashtag.subset, "HTO-4", max.cutoff = 'q99', pt.size = 0.01) +
  labs(x = "t-SNE1", y = "t-SNE2") + theme_void() + theme(legend.position = "none") + ggtitle("")

ggsave(cowplot::plot_grid(p1, p2, p3, p4, nrow = 2), file = "plots/HTO_tSNE.png", dpi = 500, width = 4, height = 4)

p1 <- HTOHeatmap(pbmc.hashtag, assay = "HTO", ncells = 5000) +
  scale_fill_gradientn(colors = jdb_palette("solar_extra")) +
  theme(legend.position = "none")
ggsave(p1, file = "../plots/visualize_HTO_heatmap.png", width = 3, height = 1.8)

# Now visualize on the cellranger map
df <- data.frame(pbmc.hashtag@meta.data)
df$Barcode <- paste0(rownames(df), "-3")
#write.table(df, file = "hash_assignments.tsv", sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE)
